yarn: switch package manager from npm to yarn
yarn is a much better package manager than npm overall, and its linker
is far more sophisticated than the one from npm. Unfortunately, the
linker in question is dubbed as the "Plug n' Play feature" of yarn, and
MathJaX isn't quite ready to move to it. In this commit, we switch to
yarn berry, preserving the node_modules directory, to see if it piques
interest of the community to switch to the yarn linker.