ui/safe: parse link protocol containing newlines correctly
If a href protocol contains `\n` or `\r` characters before `:`,
the ui/safe newline parsing will interpret the protocol as `''`
and allow it. However, the browser will ignore the newline
characters and interpret the protocol as the full word, allowing
users to submit MathJax links with `javascript:` hrefs.

This commit updates the protocol-parsing regex to include new line
characters, just ass the browser does.